refactor: 配置键名规范化、ExcelConfig 集成、Repository 模式
- 配置键名统一规范化 (enable→enabled, expire→ttl, name→connection 等)
- 全面使用 ExcelConfig 值对象替代原始数组访问
- 新增 HyperfExcelLogRepository 实现 ExcelLogRepositoryInterface
- HyperfRedisAdapter 委托 PhpRedisAdapter 统一 Redis 操作
- CleanFileProcess 使用 ExcelConfig 替代原始数组
- excel_log 表 id 字段改为 bigint unsigned
- README 补充从 vartruexuan/hyperf-excel 迁移指南

Made-with: Cursor

+++ src/migrations/excel_log.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `excel_log` (
-     `id` int unsigned NOT NULL AUTO_INCREMENT,
+     `id` bigint unsigned NOT NULL AUTO_INCREMENT,
      `token` varchar(64) NOT NULL DEFAULT '',
      `type` enum('export','import') NOT NULL DEFAULT 'export' COMMENT '类型:export导出import导入',
      `config_class` varchar(250) NOT NULL DEFAULT '',
